backend/requirements.txt
alembic==1.8.1
aniso8601==9.0.1
appdirs==1.4.4
asgiref==3.5.2
astroid==2.12.9
Babel==2.10.3
certifi==2022.6.15.1
chardet==3.0.4
click==8.1.3
cloudinary==1.29.0
colorama==0.4.5
distlib==0.3.6
dj-database-url==1.0.0
Django==4.1.1
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-grappelli==3.0.3
django-heroku==0.3.1
djangorestframework==3.13.1
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-Moment==1.0.4
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.3
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
Mako==1.2.2
MarkupSafe==2.1.1
mccabe==0.7.0
packaging==21.3
Pillow==9.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pylint==2.15.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.41
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.12
virtualenv==20.16.5
Werkzeug==2.2.2
whitenoise==6.2.0
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.1
